pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[vertexai]>=1.0.0",
    "litellm>=1.70.0",
]
